requirements.txt

python-dotenv==1.0.1
psycopg2-binary
SQLAlchemy==2.0.45
streamlit==1.52.2

